% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{calc_interaction_pvalues}
\alias{calc_interaction_pvalues}
\title{Calculate interaction p-values}
\usage{
calc_interaction_pvalues(effect1, effect2, se1, se2, pvals_only = TRUE)
}
\arguments{
\item{effect1}{numeric vector}

\item{effect2}{numeric vector}

\item{se1}{numeric vector}

\item{se2}{numeric vector}

\item{pvals_only}{logical. Indicate whether to return only interaction
p-values, or a dataframe also including the input betas and standard
errors. Default = \code{TRUE}}
}
\value{
df or numeric vector of interaction p-values (determined by
\code{pvals_only} argument)
}
\description{
Calculate interaction p-values
}
