% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{capitalise_first_letter_var_labels}
\alias{capitalise_first_letter_var_labels}
\title{Update variable labels - capitalise the first letter}
\usage{
capitalise_first_letter_var_labels(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
Update variable labels - capitalise the first letter
}
\seealso{
Other Labelled data utility functions: 
\code{\link{str_remove_from_var_labels}()},
\code{\link{str_replace_all_var_labels}()}
}
\concept{Labelled data utility functions}
