% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{logical_to_labelled}
\alias{logical_to_labelled}
\title{Convert columns of type logical to numeric labelled 'yes'/'no'}
\usage{
logical_to_labelled(df, selected_cols = NULL)
}
\arguments{
\item{df}{data frame}

\item{selected_cols}{character. A vector of selected column names to convert
to labelled numeric. These should all be of type logical.}
}
\value{
A data frame
}
\description{
Values in logical columns are converted from \code{TRUE}/\code{FALSE} to
\code{1}/\code{2}, with labels 'Yes'/'No'.
}
