% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummies.R
\name{make_dummy_df}
\alias{make_dummy_df}
\title{Make dummy data frame}
\usage{
make_dummy_df()
}
\value{
df
}
\description{
Returns a dataframe with 6 rows and columns of types character, integer and factor
}
\examples{
make_dummy_df()

str(make_dummy_df())
}
