% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{my_skim}
\alias{my_skim}
\title{Functions for summarising tabular data.
Quickly summarise a dataframe}
\usage{
my_skim(data, ...)
}
\arguments{
\item{data}{A tibble, or an object that can be coerced into a tibble.}

\item{...}{Columns to select for skimming. When none are provided, the
default is to skim all columns.}
}
\description{
A tidyverse-friendly summary function that summarises a dataframe by column
type.
}
\details{
Works with \code{dplyr::group_by()} and the pipe. See the
\code{\link[skimr]{skim}} documentation for more details. Adapts the
\code{skimr::skim()} function to include proportion counts for factor
variables
}
\examples{
library(magrittr)
# summarise the iris dataset
my_skim(iris)

# summarise the mtcars dataset by transmissions type ("am": 0 = automatic, 1 = manual)
mtcars \%>\%
  dplyr::mutate(dplyr::across(tidyselect::all_of(c("cyl", "vs", "am", "gear", "carb")), as.factor))
  dplyr::group_by(am) \%>\%
  my_skim()

}
