% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(
  df,
  x,
  y,
  fill,
  label = NULL,
  x_order = NULL,
  y_order = NULL,
  xlab = "x title",
  ylab = "y title",
  legend_lab = "legend title",
  xlab_text_angle = 90,
  ylab_text_angle = 0,
  xlab_text_size = NULL,
  ylab_text_size = NULL,
  geom_tile_color = "white"
)
}
\arguments{
\item{df}{df}

\item{x}{character}

\item{y}{character}

\item{fill}{character}

\item{x_order}{chracter vector}

\item{y_order}{character vetor}

\item{xlab}{character}

\item{ylab}{character}

\item{legend_lab}{character}

\item{xlab_text_angle}{numeric}

\item{ylab_text_angle}{numeric}

\item{xlab_text_size}{numeric}

\item{ylab_text_size}{numeric}

\item{geom_tile_color}{character}
}
\value{
ggplot object
}
\description{
Plots a heatmap using geom_tile()
}
