% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_df_as_call_to_tibble}
\alias{print_df_as_call_to_tibble}
\title{Print a data frame to the console for copying and pasting to a call to \code{tibble()}}
\usage{
print_df_as_call_to_tibble(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
NULL.
}
\description{
Convenience function
}
\examples{
print_df_as_call_to_tibble(head(iris))
}
