% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{revalue_col}
\alias{revalue_col}
\title{Revalue values in a dataframe column}
\usage{
revalue_col(df, colname, dict, default_value = NULL)
}
\arguments{
\item{df}{dataframe.}

\item{colname}{character. Name of column to be relabelled}

\item{dict}{a named vector. \code{names(dict)} are the 'keys' i.e. the
existing values in \code{df[[colname]]} to be replaced. These should be
unique. An error is raised if non-unique values are found in
\code{names(dict)}}

\item{default_value}{default value to use for values in \code{df[[colname]]}
that are not present in \code{names(dict)}. By default this is \code{NULL},
meaning that values not present in \code{names(dict)} will remain
unchanged.}
}
\value{
dataframe
}
\description{
Similar idea to
\href{https://pandas.pydata.org/pandas-docs/stable/reference/api/pandas.DataFrame.replace.html}{pandas.replace
method} using a dictionary for the \code{value} argument. \strong{Note:} only works with columns that are of type numeric (including integer) or character.
}
