% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelled.R
\name{str_remove_from_var_labels}
\alias{str_remove_from_var_labels}
\title{Update variable labels - remove a string from all variable labels}
\usage{
str_remove_from_var_labels(df, pattern = "\\\\s\\\\(f[:digit:]+.*\\\\)$")
}
\arguments{
\item{df}{A data frame}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:about_search_regex]{stringi::stringi-search-regex}. Control options with
\code{\link[stringr:modifiers]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:modifiers]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:modifiers]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:modifiers]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}
}
\value{
A data frame
}
\description{
Update variable labels - remove a string from all variable labels
}
\seealso{
Other Labelled data utility functions: 
\code{\link{capitalise_first_letter_var_labels}()},
\code{\link{str_replace_all_var_labels}()}
}
\concept{Labelled data utility functions}
