% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap}
\usage{
plot_heatmap(
  df,
  x,
  y,
  fill,
  label = NULL,
  label_size = TRUE,
  x_order = NULL,
  y_order = NULL,
  xlab = "x title",
  ylab = "y title",
  legend_lab = "legend title",
  xlab_text_angle = 90,
  ylab_text_angle = 0,
  xlab_text_size = NULL,
  ylab_text_size = NULL,
  xtitle_text_size = NULL,
  ytitle_text_size = NULL,
  legend_title_size = NULL,
  legend_text_size = NULL,
  geom_tile_color = "white",
  plotly_plot = FALSE
)
}
\arguments{
\item{df}{df}

\item{x}{character}

\item{y}{character}

\item{fill}{character}

\item{label}{character. Annotate tiles}

\item{label_size}{numeric. Text size for tile annotations}

\item{x_order}{chracter vector.}

\item{y_order}{character vetor}

\item{xlab}{character}

\item{ylab}{character}

\item{legend_lab}{character}

\item{xlab_text_angle}{numeric}

\item{ylab_text_angle}{numeric}

\item{xlab_text_size}{numeric}

\item{ylab_text_size}{numeric}

\item{xtitle_text_size}{numeric}

\item{ytitle_text_size}{numeric}

\item{legend_title_size}{numeric}

\item{legend_text_size}{numeric}

\item{geom_tile_color}{character}

\item{plotly_plot}{logical. Set as \code{TRUE} if wrapping this plot with
\code{\link[plotly]{ggplotly}}. Default is \code{FALSE}.}
}
\value{
ggplot object
}
\description{
Plots a heatmap using \code{\link[ggplot2]{geom_tile}}.
}
\details{
TODO: read this \href{https://jcoliver.github.io/learn-r/006-heatmaps.html}{blogpost}
}
